moved colorassigner to hooks
